+++ auth0/src/main/java/com/auth0/android/authentication/ParameterBuilder.java
@@ -53,6 +53,7 @@ public class ParameterBuilder {
     public static final String GRANT_TYPE_PASSWORD_REALM = "http://auth0.com/oauth/grant-type/password-realm";
     public static final String GRANT_TYPE_JWT = "urn:ietf:params:oauth:grant-type:jwt-bearer";
     public static final String GRANT_TYPE_AUTHORIZATION_CODE = "authorization_code";
+    public static final String GRANT_TYPE_MFA_OTP = "http://auth0.com/oauth/grant-type/mfa-otp";
 
     public static final String SCOPE_OPENID = "openid";
     public static final String SCOPE_OFFLINE_ACCESS = "openid offline_access";
@@ -79,7 +80,7 @@ public class ParameterBuilder {
     /**
      * Sets the 'client_id' parameter
      *
-     * @param clientId clientID
+     * @param clientId the application's client id
      * @return itself
      */
     public ParameterBuilder setClientId(String clientId) {